Issue #48 Add bitshift operation comparison of time for movement slowing. -e Original/build commit ID: 6fb010ffe7e07546af2f7b65683c16c84cc45274

+++ game_state.c
@@ -70,7 +70,7 @@ typedef struct {
     // Inventory items
     UINT8 inventory[7U];
     
-    unsigned int last_movement_time;
+    UINT16 last_movement_time;
 } game_state_t;
